+++ src/app/month-calendar/month-calendar.component.spec.ts
@@ -3,8 +3,10 @@ import {MonthCalendarComponent} from './month-calendar.component';
 import {UtilsService} from '../common/services/utils/utils.service';
 import {CalendarNavComponent} from '../calendar-nav/calendar-nav.component';
 import {MonthCalendarService} from './month-calendar.service';
-import {IMonth} from './month.model';
+import * as momentNs from 'moment';
 import {Moment} from 'jalali-moment';
+import {IMonth} from './month.model';
+const moment = momentNs;
 
 describe('Component: MonthCalendarComponent', () => {
   let component: MonthCalendarComponent;
@@ -36,7 +38,7 @@ describe('Component: MonthCalendarComponent', () => {
       disabled: false,
       text: ''
     };
-    const defaultCssClasses: { [klass: string]: boolean } = {
+    const defaultCssClasses: {[klass: string]: boolean} = {
       'dp-selected': false,
       'dp-current-month': false
     };
@@ -70,6 +72,12 @@ describe('Component: MonthCalendarComponent', () => {
         ...defaultCssClasses,
         'custom-class': true
       });
+    });
+
+    it('should emit event goToCurrent function called', () => {
+      spyOn(component.onGoToCurrent, 'emit');
+      component.goToCurrent();
+      expect(component.onGoToCurrent.emit).toHaveBeenCalledWith();
     });
   });
 });